refactor: 구현체를 참조하고 있던 부분을 인터페이스를 참조하도록 변경함.
